# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Spotdesking API'
  description: 'Spotdesking [Aka QRCLEAN] API Description'
  version: 1.0.0
servers:
  -
    url: /api
paths:
  /auth/login:
    post:
      tags:
        - Autorizacion
      summary: 'Inicia sesion para el usuario y devuelve el token de acceso'
      operationId: cdf452d2ed99a23c8464e289c7fce73b
      requestBody:
        content:
          application/json:
            schema:
              required:
                - username
                - password
              properties:
                username:
                  description: username
                  type: string
                password:
                  description: password
                  type: string
                remember_me:
                  description: remember_me
                  type: boolean
              type: object
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                    result:
                      description: 'Resultado de la operacion'
                      type: string
                      enum:
                      - ok
                    timestamp:
                      description: 'Fecha de ejecucion de la accion'
                      type: string
                      format: date-time
                    access_token: 
                      type: string
                    token_type: 
                      type: string
                    expires_at: 
                      type: string
                type: object
        400:
          description: 'Error en la peticion'
          content:
            application/json:
              schema:
                properties:
                  result:
                    description: 'Resultado de la operacion'
                    type: string
                    enum:
                    - error
                  error:
                    description: 'Descripcion del error'
                    type: string
                  timestamp:
                    description: 'Fecha de ejecucion de la accion'
                    type: string
                    format: date-time
                type: object
        401:
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                properties:
                  result:
                    description: 'Resultado de la operacion'
                    type: string
                    enum:
                    - error
                  error:
                    description: 'Descripcion del error'
                    type: string
                  timestamp:
                    description: 'Fecha de ejecucion de la accion'
                    type: string
                    format: date-time
                type: object
      security:
        -
          passport: []
  /auth/logout:
    get:
      tags:
        - Autorizacion
      summary: 'Cerrar sesion para el usuario actual'
      operationId: d7b22dfe8ef07f2fa0268b7b6349663a
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  result:
                    description: 'Resultado de la operacion'
                    type: string
                    enum:
                    - ok
                  timestamp:
                    description: 'Fecha de ejecucion de la accion'
                    type: string
                    format: date-time
                  message:
                    description: 'Mensaje de resultado de la operacion'
                    type: string
                type: object
        400:
          description: 'Error en la peticion'
          content:
            application/json:
              schema:
                properties:
                  result:
                    description: 'Resultado de la operacion'
                    type: string
                    enum:
                    - error
                  error:
                    description: 'Descripcion del error'
                    type: string
                  timestamp:
                    description: 'Fecha de ejecucion de la accion'
                    type: string
                    format: date-time
                type: object
      security:
        -
          passport: []
  /auth/user:
    get:
      tags:
        - Autorizacion
      summary: 'Obtiene los datos del usuario actual'
      operationId: 04ff6d178635c85a0486987fea31407d
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  result:
                    description: 'Resultado de la operacion'
                    type: string
                    enum:
                    - ok
                  timestamp:
                    description: 'Fecha de ejecucion de la accion'
                    type: string
                    format: date-time
                  id: 
                    type: integer
                    format: int32
                  name: 
                    type: string
                  email: 
                    type: string
                  email_verified_at: 
                    type: string
                  created_at: 
                    type: string
                  updated_at: 
                    type: string
                  img_usuario: 
                    type: string
                  cod_nivel: 
                    type: integer
                    format: int32
                  theme: 
                    type: string
                    format: nullable
                  collapse: 
                    type: integer
                    format: int32
                  last_login: 
                    type: string
                  nivel_acceso: 
                    type: integer
                    format: int32
                  id_cliente: 
                    type: integer
                    format: int32
                  val_timezone: 
                    type: string
                  token_acceso: 
                    type: string
                  def_camera: 
                    type: string
                  email_expire_at: 
                    type: string
                    format: nullable
                  id_usuario_supervisor: 
                    type: string
                    format: nullable
                  val_vista_puestos: 
                    type: string
                  token_expires: 
                    type: integer
                    format: int32
                  id_externo: 
                    type: string
                type: object
        400:
          description: 'Error en la peticion'
          content:
            application/json:
              schema:
                properties:
                  result:
                    description: 'Resultado de la operacion'
                    type: string
                    enum:
                    - error
                  error:
                    description: 'Descripcion del error'
                    type: string
                  timestamp:
                    description: 'Fecha de ejecucion de la accion'
                    type: string
                    format: date-time
                type: object
      security:
        -
          passport: []
...
components:
  securitySchemes:
    passport:
      type: oauth2
      description: 'Laravel passport oauth2 security.'
      in: header
      scheme: https
      flows:
        password:
          authorizationUrl: '/api/auth/login'
          tokenUrl: '/api/auth/login'
          refreshUrl: '/api/auth/token/refresh'
          scopes: []